feat(cdp): segment follow up pr (#32651)
