+++ frontend/src/scenes/hog-functions/mapping/HogFunctionMappings.tsx
@@ -66,9 +66,11 @@ const MappingSummary = memo(function MappingSummary({
             </span>
             <IconArrowRight className="text-secondary" />
             <span>
-                {typeof firstInputValue === 'object'
+                {mapping.name
+                    ? humanize(mapping.name)
+                    : typeof firstInputValue === 'object'
                     ? JSON.stringify(firstInputValue)
-                    : humanize(mapping.name ?? firstInputValue)}
+                    : humanize(firstInputValue)}
             </span>
             <span className="flex-1" />
             {mapping.disabled ? <LemonTag type="danger">Disabled</LemonTag> : null}
