+++ plugin-server/src/utils/try-catch.ts
@@ -1,0 +1,24 @@
+/**
+ * Sometimes you want to try catch a simple function in a single assignment and this function helps you do that
+ * So instead of:
+ *
+ * let result: T
+ * try {
+ *     result = await fn()
+ * } catch (e) {
+ *     result = e
+ * }
+ *
+ * you can do:
+ *
+ * const [error, result] = await tryCatch(fn)
+ */
+
+export async function tryCatch<T>(fn: () => Promise<T>): Promise<[Error | null, T | null]> {
+    try {
+        const result = await fn()
+        return [null, result]
+    } catch (e) {
+        return [e, null]
+    }
+}
